bump(k8s.io/kubernetes): 4a65fa1f35e98ae96785836d99bf4ec7712ab682
Also includes transitive-only bumps/drops:
bump(github.com/appc/spec/schema): 818ac4d0073424f4e0a46f45abaa147ccc1b5a24
bump(github.com/aws/aws-sdk-go): 4944a94a3b092d1dad8a964415700a70f55e580a
bump(github.com/coreos/rkt): 71d7331e46f20aaa41bed05e861bfeca92249968
bump(github.com/google/cadvisor): a8085bf9276c22f16dbcd7aa56f0d4d0626a0b2e
bump(github.com/juju/ratelimit): 77ed1c8a01217656d2080ad51981f6e99adaa177
bump(github.com/mesos/mesos-go): ed907b10717e66325cf2894eb90a0553a89fcb11
bump(google.golang.org/api): 0c2979aeaa5b573e60d3ddffe5ce8dca8df309bd
bump(google.golang.org/grpc): 4bd040ce23a624ff9a1d07b0e729ee189bddd51c
drop(github.com/prometheus/client_golang/extraction)
drop(github.com/jlhawn/go-crypto): cd738dde20f0b3782516181b0866c9bb9db47401
bump(golang.org/x/net): c2528b2dd8352441850638a8bb678c2ad056fd3e

+++ Godeps/_workspace/src/k8s.io/kubernetes/test/e2e/framework.go
@@ -20,12 +20,12 @@ import (
 	"bytes"
 	"fmt"
 	"strings"
+	"sync"
 	"time"
 
 	"k8s.io/kubernetes/pkg/api"
 	client "k8s.io/kubernetes/pkg/client/unversioned"
 	"k8s.io/kubernetes/pkg/fields"
-	"k8s.io/kubernetes/pkg/labels"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
@@ -40,22 +40,23 @@ type Framework struct {
 	Client                   *client.Client
 	NamespaceDeletionTimeout time.Duration
 
-	// Allows to override the initialization of the namespace
-	nsCreateFunc func(string, *client.Client) (*api.Namespace, error)
+	gatherer containerResourceGatherer
+	// Constraints that passed to a check which is exectued after data is gathered to
+	// see if 99% of results are within acceptable bounds. It as to be injected in the test,
+	// as expectations vary greatly. Constraints are groupped by the container names.
+	addonResourceConstraints map[string]resourceConstraint
+
+	logsSizeWaitGroup    sync.WaitGroup
+	logsSizeCloseChannel chan bool
+	logsSizeVerifier     *LogsSizeVerifier
 }
 
 // NewFramework makes a new framework and sets up a BeforeEach/AfterEach for
 // you (you can write additional before/after each functions).
 func NewFramework(baseName string) *Framework {
-	return InitializeFramework(baseName, createTestingNS)
-}
-
-// InitializeFramework initialize the framework by allowing to pass a custom
-// namespace creation function.
-func InitializeFramework(baseName string, nsCreateFunc func(string, *client.Client) (*api.Namespace, error)) *Framework {
 	f := &Framework{
-		BaseName:     baseName,
-		nsCreateFunc: nsCreateFunc,
+		BaseName:                 baseName,
+		addonResourceConstraints: make(map[string]resourceConstraint),
 	}
 
 	BeforeEach(f.beforeEach)
@@ -73,7 +74,7 @@ func (f *Framework) beforeEach() {
 	f.Client = c
 
 	By("Building a namespace api object")
-	namespace, err := f.nsCreateFunc(f.BaseName, f.Client)
+	namespace, err := createTestingNS(f.BaseName, f.Client)
 	Expect(err).NotTo(HaveOccurred())
 
 	f.Namespace = namespace
@@ -85,6 +86,21 @@ func (f *Framework) beforeEach() {
 	} else {
 		Logf("Skipping waiting for service account")
 	}
+
+	if testContext.GatherKubeSystemResourceUsageData {
+		f.gatherer.startGatheringData(c, time.Minute)
+	}
+
+	if testContext.GatherLogsSizes {
+		f.logsSizeWaitGroup = sync.WaitGroup{}
+		f.logsSizeWaitGroup.Add(1)
+		f.logsSizeCloseChannel = make(chan bool)
+		f.logsSizeVerifier = NewLogsVerifier(c, f.logsSizeCloseChannel)
+		go func() {
+			f.logsSizeVerifier.Run()
+			f.logsSizeWaitGroup.Done()
+		}()
+	}
 }
 
 // afterEach deletes the namespace, after reading its events.
@@ -92,7 +108,7 @@ func (f *Framework) afterEach() {
 	// Print events if the test failed.
 	if CurrentGinkgoTestDescription().Failed {
 		By(fmt.Sprintf("Collecting events from namespace %q.", f.Namespace.Name))
-		events, err := f.Client.Events(f.Namespace.Name).List(labels.Everything(), fields.Everything())
+		events, err := f.Client.Events(f.Namespace.Name).List(api.ListOptions{})
 		Expect(err).NotTo(HaveOccurred())
 
 		for _, e := range events.Items {
@@ -103,6 +119,8 @@ func (f *Framework) afterEach() {
 		// you may or may not see the killing/deletion/cleanup events.
 
 		dumpAllPodInfo(f.Client)
+
+		dumpAllNodeInfo(f.Client)
 	}
 
 	// Check whether all nodes are ready after the test.
@@ -110,14 +128,27 @@ func (f *Framework) afterEach() {
 		Failf("All nodes should be ready after test, %v", err)
 	}
 
-	By(fmt.Sprintf("Destroying namespace %q for this suite.", f.Namespace.Name))
+	if testContext.DeleteNamespace {
+		By(fmt.Sprintf("Destroying namespace %q for this suite.", f.Namespace.Name))
 
-	timeout := 5 * time.Minute
-	if f.NamespaceDeletionTimeout != 0 {
-		timeout = f.NamespaceDeletionTimeout
+		timeout := 5 * time.Minute
+		if f.NamespaceDeletionTimeout != 0 {
+			timeout = f.NamespaceDeletionTimeout
+		}
+		if err := deleteNS(f.Client, f.Namespace.Name, timeout); err != nil {
+			Failf("Couldn't delete ns %q: %s", f.Namespace.Name, err)
+		}
+	} else {
+		Logf("Found DeleteNamespace=false, skipping namespace deletion!")
 	}
-	if err := deleteNS(f.Client, f.Namespace.Name, timeout); err != nil {
-		Failf("Couldn't delete ns %q: %s", f.Namespace.Name, err)
+
+	if testContext.GatherKubeSystemResourceUsageData {
+		f.gatherer.stopAndPrintData([]int{50, 90, 99, 100}, f.addonResourceConstraints)
+	}
+
+	if testContext.GatherLogsSizes {
+		close(f.logsSizeCloseChannel)
+		f.logsSizeWaitGroup.Wait()
 	}
 	// Paranoia-- prevent reuse!
 	f.Namespace = nil
@@ -127,6 +158,12 @@ func (f *Framework) afterEach() {
 // WaitForPodRunning waits for the pod to run in the namespace.
 func (f *Framework) WaitForPodRunning(podName string) error {
 	return waitForPodRunningInNamespace(f.Client, podName, f.Namespace.Name)
+}
+
+// WaitForPodRunningSlow waits for the pod to run in the namespace.
+// It has a longer timeout then WaitForPodRunning (util.slowPodStartTimeout).
+func (f *Framework) WaitForPodRunningSlow(podName string) error {
+	return waitForPodRunningInNamespaceSlow(f.Client, podName, f.Namespace.Name)
 }
 
 // Runs the given pod and verifies that the output of exact container matches the desired output.
@@ -145,7 +182,7 @@ func (f *Framework) WaitForAnEndpoint(serviceName string) error {
 	for {
 		// TODO: Endpoints client should take a field selector so we
 		// don't have to list everything.
-		list, err := f.Client.Endpoints(f.Namespace.Name).List(labels.Everything())
+		list, err := f.Client.Endpoints(f.Namespace.Name).List(api.ListOptions{})
 		if err != nil {
 			return err
 		}
@@ -160,11 +197,11 @@ func (f *Framework) WaitForAnEndpoint(serviceName string) error {
 			}
 		}
 
-		w, err := f.Client.Endpoints(f.Namespace.Name).Watch(
-			labels.Everything(),
-			fields.Set{"metadata.name": serviceName}.AsSelector(),
-			rv,
-		)
+		options := api.ListOptions{
+			FieldSelector:   fields.Set{"metadata.name": serviceName}.AsSelector(),
+			ResourceVersion: rv,
+		}
+		w, err := f.Client.Endpoints(f.Namespace.Name).Watch(options)
 		if err != nil {
 			return err
 		}
